Add semantic HTML5 elements to index headings
W3C's HTML5 specification instructs a standard way to create
subheadings in "4.12.1 Subheadings, subtitles, alternative titles and
taglines". [1]

A `<h2>` element is not appropriate here as a subheading, because the
subheading "Roppa interior~" is not part of the document outline but a
Catalonian translation of the `<h1>` heading.

Adding a `lang="ca"` attribute to the Catalonian subheading increases
screenreader compatibility and semantics even further. Screenreaders
will then use Catalonian pronounction instead of English if supported.

[1]: http://www.w3.org/html/wg/drafts/html/CR/common-idioms.html#sub-head