Rework generate_name() logic
Adds a new config option POMF_FILES_LENGTH. This will allow the user to
adjust the length of filenames. Defaults to 6.

The filename generator never used character z in new filenames. There
are 27 characters in a-z, but the old code uses a magic number of 26
characters. This bug was introduced in commit
5b77a4eaab418836820b2a0bad7c88f0b88136a0.

Instead of relying on magic numbers like with mt_rand(), conveniently
use arrays with array_rand() instead. More characters can be added with
array_merge() in future if desired.