Indent and format project to PSR-2 style
PSR-2 [1] is a highly regarded style guide in the PHP community
according to Stack Overflow [2].

The indentation was out of place in the source code. I've seen a
combination of 1 space tabs, 3 space tabs, 4 space tabs and 5 space
tabs. Some closing braces on the same line and a lot of the code
(especially moe code) had reached the point of losing readability.

Converted with php-cs-fixer and some hand adjustment.

[1]: http://www.php-fig.org/psr/psr-2/
[2]: https://stackoverflow.com/questions/3638083/list-of-highly-regarded
-php-style-guides