Remove moe panel from core
The moe panel is more of an extension of the core, but not necessary for
Pomf's functionality. The code also suffers from bad practices and it's
more of a prototype. The panel is also mostly unmaintained by its
author.

This code will be in a seperate Git repository called pomf-moe from now
on.