Remove moe panel tables from core database schema
Accounts, invites and reports are not used in the core. These will be
moved to pomf-moe repository. You don't need it anymore.

While the change of `orginalname` to `originalname` is good to mention,
it is better served in the documentantion for those who are upgrading
instead of laying around useless here in this code repository.

OK pantsucat