Remove <header> element from index pages
The <header> element represents a group of introductory or navigational
aids. It is semantically incorrect to use it with only one element
inside.

The <header> element was introduced in commit 87014fa08f. That commit
describes the correct usage of <header> element with multiple elements
inside <header>.

This bug was introduced in commit fd566d3d7f.