Replace #upload-btn <a> element with <button>
Corrects semantics and increases accessibility.

The <a> element represents a hyperlink. A hyperlink is "a link from a
hypertext file or document to another location or file." This is not
what #upload-btn is, it's a button.

Ask yourself: "Does the element change the URL?" In our case, the upload
button does not so it is incorrect to use an <a> element here.

The idea to use <a> elements as buttons comes from years ago when you
couldn't style <button> elements in legacy browsers. Today you can, so
there's no reason to break the semantics.

Reference: http://blogs.ancestry.com/techroots/buttons-vs-anchors/