Change virus scan banner class to alert-info
alert-info conveys a message "this needs your attention, but it's not
super important."

alert-error conveys a message "you screwed up, please correct the error
or leave the site."

This means alert-error should only be used for errors the user has made.
Using it for an informational message about virus scans is incorrect.