Fix README.md grammar "an attacker"
English nouns begin with an indefinite article "a" or "an". Usually "a"
is used when the word that follows starts with a consontant, and "an"
is used when the word that follows starts with a vowel.

The word "attacker" starts with a vowel and has no different
pronunciation of the first letter, so the correct indefinite article to
use is "an". The correct way to pronounce it is "an attacker".

This fixes an error introduced in commit db66648.