Wrap CSV response values in double quotes (")
While double quotes are mostly optional in implementations, some CSV
linters want double quotes for whatever the reason. Assumably, it
reduces the amount of bugs and avoids possible parsing errors. This also
has very little relevance for not trusting the user input, but better be
safe than sorry.

This commit is not very pretty. In a future commit, this issue will be
addressed with arrays and fputcsv().