Fix incorrect progress/load event listeners
This is a symptom of copy-paste code. A file would be uploaded, but it
would return the "Something went wrong" error message without telling
why.

Me, ewhal@ and bursey on IRC tried had a look to see what was wrong, but
none of us could catch this subtle bug. It wasn't until alchimist@
(cenci0) looked at this and found the issue. Thanks.

This bug was introduced in commit 2751811.