Remove grunt-contrib-imagemin image minification
Ever since liberating images in c76f70c, there's no really a need to
minify images. Furthermore, the dependency update in 6ef710a changed PNG
minification to be lossless, so most of the benefits were lost (because
our images are already compressed in the repository).

We don't need this anymore.