Remove Catalan pages/translations
This is not the right way to do i18n. Banners are difficult and buggy
with translations, the folder structure is wrong and the result is
duplicated markup.

For the sake of maintainability, I am liberally removing this from core.
Future i18n attempts should use real gettext implementations and
possibly seperate i18n files from core as an extension.

Because Cupkle contributed the Catalan translation, this also removes
him/her from contributors.