Rewrite the FAQ page with better text and semantics
The previous FAQ is difficult to read and doesn't actually answer to
common questions. It also lacks semantics.

Because the previous FAQ had reached a point of being next to useless
and unfixable with minor edits, it had to be rewritten.

This commit addresses the most commonly asked questions. It uses HTML5
headings for addresssing each question for proper document outline. For
future improvement, more may be added as necessary.