FAQ: Wrap body in <article> element
The <article> element is an element introduced in HTML5 for semantic
sectioning.

faq.swig is missing this semantic element, so this commit adds it.

It doesn't seem appropriate to use <article> in layout.swig, because
it's used for every page, also ones that are not articles.

Reference: http://www.w3.org/TR/html5/sections.html#the-article-element