pages: Move nav.swig includes from jumbotron to bottom
nav.swig includes were inconsistently at either bottom of the page or
inside the jumbotron. For FAQ, the navigation was also nested
incorrectly inside the <article> element, which means the navigation
incorrectly assumed to be in-article navigation (table of contents).

The intention is to have these nav.swig (site-wide navigation) includes
outside of <article> scope. Future improvement is desired to have all
pages like other.swig and tools.swig to have <article> elements.