FAQ: Clarify "valid notice" with Copyright Act 1968
A "valid notice" for copyright takedown is questionable and doesn't tell
much of a clue what needs to be in a takedown notice. Change it to
specify Copyright Act 1968 (Australia), which is the current copyright
legislation in Australia.

For future improvement, a copyright takedown page could be created with
more detailed instructions what needs to be included in a copyright
takedown notice.