Revert default max-file-size from 120MiB to 10MiB
ewhal@ didn't understand how changing the maximum file size works, so he
edited the JavaScript code directly to set the default to his desired
120MiB.

For all clones, this may not be viable and it's not a sane default in my
opinion. This doesn't change the fact that the maximum file size remains
to be what it's configured in the configuration (Gruntfile.js).

This bug was introduced in commit ec0154b, and potentially others.

Vague OK from ewhal@ many weeks ago.