Revert <input type="hidden"> in upload_form_nojs.swig
Hidden <input> values did not work as I expected for POST requests. In
GET requests, the hidden values are added as query parameters to the
URL. For POST requests, the hidden inputs are simply more form-data to
upload in the POST request (another "file").

Because of this, the HTML output parameter is never passed to upload.php
and so it returns the default response instead (JSON). The goal is to
use the HTML response.

This reverts commit d6dc103c7229cb0e2e177ce0dfc0cfb528025ca9, where the
bug was introduced.

OK ewhal@