Fix generate_name() imaginary column "pomf" to "files"
(Partially reverts commit e0bb4e0.)

Actually fix generate_name() COUNT query this time. It still had an
imaginary table "pomf" from e0bb4e0. Change it to "files" like in the
schema.

For code consistency, :filename is reverted back to :name as it was
before commit e0bb4e0. That commit is bogus and doesn't fix the issue
due to skipping the code review process.

The problem with this COUNT query in first place was that it checked
imaginary columns (name) from an imaginary table (pomf) in the database.
This led to the "same filename" check not working at all, and the other
code only relying on checking the same hash (currently SHA-1) for
returning the filename before generate_name().

By luck, generated filenames could collide and another file of the same
generated filename would be returned, discarding the new upload.

I discovered this bug about a month ago myself, but back then there was
no reliable way to reproduce it and no good description of the bug.

This bug was introduced in commit 5b77a4e.

Reported-by: Vhum Vhum <vhumvhum1@gmail.com>
Suggested-by: Luminarys <luminarys@fuwa.se>

OK ewhal@, luminarys@