Workaround undefined length TypeError in cheesesteak
this._events[evt].length would run into JavaScript scope issues in
EventEmitter.prototype.emit function for-loop.

This is an attempt to work around the issue to resolve an innocent
looking TypeError. The code seems to work without this patch, but I
can't spot if this is some dead code or know what to do with it. Due to
technical debt, some sacrifice had to be made in readability for fixing
this bug.

This bug was introduced in commit 575b78b.

For future improvement, this function and other functions in the file
should be refactored or replaced with a more efficient implementation.