Remove Peter L. and Eric J. from faq.swig authors
Peter Lejeck's (peter.lejeck@) only commit in the file is a507748, which
is pretty much just the HTML layout in the file. I have changed that, a
lot.

Eric Johansson's (neku@) commits in the file covered the text itself.
This too has been replaced in commit 7d61e04, which rewrote all the
text in the page.

In other words, these two authors no longer have any meaningful content
in the file to be attributed. Much of the copyright goes to the
Pantsu.cat developers - which covers myself too.

OK luminarys@, "no complaints"