Remove unused page specific CSS (old, ded)
The tools page has only had alive and maintained tools on it since Pomf
2.0.0. Ded and old tools have been removed in commit
627937d0fa632a677f278af467886915c0a36f35.

It is anticipated that these style rules will be never used again. If we
did, we can add them back later: "you aren't gonna need it" (YAGNI). If
it's dead, you should probably remove the tool from the tools page
instead of beating a dead horse.

[wub@partyvan.eu: Modified Git commit message, as discussed]
Signed-off-by: Juuso Lapinlampi <wub@partyvan.eu>