Deprecate Swig banners
I have a belief that Swig banners are a Grunt only thing, and they don't
really integrate very well into Swig anyway. It's kind of a hack, and a
bad one because we don't need it anymore.

By deprecating Swig banners, we should have easier time replacing
Grunt/npm with Make in a future release.

Reviewed-by: Eliot Whalan <ewhal@pantsu.cat>