pages: Add metadata generator value to <head>
WHATWG's HTML standard vaguely implies in section "4.2.5.1 Standard
metadata names" that documents that lack a metadata generator value were
written by a user in a text editor. This is not the case for Pomf.

The next Pomf release is expected to be version 2.1.0, which was picked
here for the value.

While we are here, update copyright year to 2016 for this template.