Fix/Update LibreJS @source URL to cgit tree
git.pantsu.cat has switched from gogs to cgit. gogs used a slightly
different URLs to point to tree, so the old URLs from gogs are no longer
valid.

For consideration, to ease maintenance we'll just keep pointing the URL
to the master branch and not a specific commit or maint branch.
?h=master is not really required here, it's the default branch. I am
aware that the current maint and master branches diverge heavily from
each other in terms of JavaScript, but this will be an issue that will
disappear eventually. Thus, this is the simplest fix.

A HTTP redirect will be setup for the old URL to the new one
(eventually) to fix broken links.

Reviewed-by: Eliot Whalan <ewhal@pantsu.cat>