Revert "Remove bugs URL from package.json", set email
Actually on second thought, SemVer doesn't allow us to remove
functionality without incrementing the major version number... which we
don't want to do for this.

Set the bugs email address to bugs@pantsu.cat. Hoping here bugs@ to be
an alias to ewhal@ for the time being until a real mailing list will be
reachable at that address. This fixes the issue for the time being
without removing functionality.

It is okay that we made a mistake. We don't need to increment the major
version here now, because no release was tagged and released in this
branch.

Nevermind that Pomf isn't even published on NPM packages, so this
"feature" is never seen in its current form in the npm(1) program. We'll
still respect SemVer, for what little worth it may be to other tools.

This reverts commit 310d8d9fdd33826b9a6d5c1f9b1f8aa1c373c461.

Reviewed-by: Eliot Whalan <ewhal@pantsu.cat>