Deprecate file expiration support in Pomf 2.1.0
File expiration support has never been properly implemented in Pomf core
or moe panel. It serves as legacy baggage before nekunekus got an idea
to create nokonoko/Uguu for temporary uploads.

In current plans, this non-feature will be removed in the 3.0.0 release
which is the next SemVer major version that allows to remove this.

Reviewed-by: Austin Gillmann <alucard@cuntflaps.me>