Deprecate get_sha1() in Pomf 2.1.0
There are two issues with get_sha1():

- The function name doesn't follow the PSR-2 coding style guide for
  function naming, followed in Pomf since version 2.0.0.
- SHA-1 is no longer considered resilient to collisions, versus SHA-256.

In current plans, this feature is to be replaced with getHash() or
similar function name in the next major version (Pomf 3.0.0).

For consideration for future SHA-256 hashing implementation, the MySQL
schema's hash value needs to be changed from char(40) to char(64) to
accommodate the longer SHA-256 length of 64 bytes in hexadecimal string.