upload.php: Deprecate ob_start (output buffering)
None of us can find a reason why we would need to buffer our output in a
PHP application. My best guess is this may be a relic from Pomf 1.0.0 or
Moe panel.

Unless someone finds a better reason to keep this, this feature is
useless and has negligible benefit and will be removed.

In respect for SemVer, deprecating this feature happens first in a minor
version before removal in a next major version.

Acked-by: Eliot Whalan <ewhal@pantsu.cat>