FAQ: Avoid using word "content" as a noun
According to the GNU Project, using the noun "content" to describe
publications and works of authorship "adopts an attitude you might
rather avoid: it treats them as a commodity whose purpose is to fill a
box and make money. In effect, it disparages the works themselves."

From now on, we'll refer to publications as uploaded or published works.

This bug was introduced in commit
7d61e0439893a2fd946ec4d4b8e32aaaac0da322.

See: https://www.gnu.org/philosophy/words-to-avoid.en.html#Content