pages: Add a non-breaking space before "MiB"
Fixes a typo where the unit symbol was missing a space character before
it.

To avoid automatic line breaks in text "128 MiB", a non-breaking space
character is used instead of a regular space character. If the text
doesn't fit at the end of a line it is moved in its entirety to the next
line.

Reviewed-by: Eliot Whalan <ewhal@pantsu.cat>