upload.php: Deprecate all named functions
Deprecates diverse_array(), generate_name(), refiles() and
upload_file(). These pollute the global namespace and do not follow the
PSR-2 naming convention.

The plan is to create a new file with a new class to share this common
code, leaving upload.php only with the last lines of code that it really
needs to do (calling other class functions).

Acked-by: Eliot Whalan <ewhal@pantsu.cat>