UploadException: Correct license header to CC BY 3.0+
The PHP license only applies to the PHP software itself. The choice to
use the PHP license for this was just a best guess for the license, and
a questionable move.

With further investigation, the PHP manuals are licensed under the free
Creative Commons Attribution 3.0 license or later. It doesn't say which
version of that license, so I will assume unported.

This fixes copyright licensing issues with UploadException.

See: https://secure.php.net/manual/en/copyright.php

Credit to Juuso Lapinlampi for this patch.