Fix CSS URL Rebasing Issues and Update Command Syntax
I found that the syntax used for cleancss is no longer accurate as url rebasing defaults to on in the latest version of cleancss (4.3.0).
Following the -h feedback it appears adding --skip-rebase fixes this step. I also exchanged the piped stout with the dedicated --output option that is now available.