Fix cleancss binary dependancies and URL rebasing (#96)
* Fix CSS URL Rebasing Issues and Update Command Syntax

I found that the syntax used for cleancss is no longer accurate as url rebasing defaults to on in the latest version of cleancss (4.3.0).
Following the -h feedback it appears adding --skip-rebase fixes this step. I also exchanged the piped stout with the dedicated --output option that is now available.

* Update Dependancies to include cleancss binary

It appears that the node_modules/.bin/cleancss binary is no longer supplied with the clean-css package and is now contained within the separate clean-css-cli package instead, thus the relevant package has been added to the dev dependencies.